fix: use single redis cache for now
