+++ package.json
@@ -22,11 +22,13 @@
     "dotenv": "^16.4.5",
     "ethers": "5",
     "express": "^4.19.2",
+    "object-hash": "^3.0.0",
     "redis": "^4.6.13"
   },
   "devDependencies": {
     "@types/express": "^4.17.17",
     "@types/node": "^20.3.1",
+    "@types/object-hash": "^3.0.6",
     "@typescript-eslint/eslint-plugin": "^6.0.0",
     "@typescript-eslint/parser": "^6.0.0",
     "eslint": "^8.42.0",
