+++ src/cache/utils.ts
@@ -1,0 +1,7 @@
+import hash from "object-hash";
+
+export function makeGetCacheKey(prefix: string) {
+  return function getCacheKey(keyParams: Record<string, string>) {
+    return `${prefix}:${hash(keyParams)}`;
+  };
+}
